Change: - Added parameter for choosing between allennlp and our beam search - Added prune hypothesis

+++ pg_salience_feature/summarizer.py
@@ -37,15 +37,17 @@ def summarize(input, vocab_path, model, model_config, output_path, batch_size, c
     write_constraints = []
     for out in output:
         results = out['results']
-        constraints = out['word_constraints']
-        if (output_file.parent / 'constraint.txt').exists():
-            os.remove(str((output_file.parent / 'constraint.txt')))
-        for constraint in constraints:
-            write_constraints.append(str(constraint))
+        if 'word_constraints' in out:
+            constraints = out['word_constraints']
+            if (output_file.parent / 'constraint.txt').exists():
+                os.remove(str((output_file.parent / 'constraint.txt')))
+            for constraint in constraints:
+                write_constraints.append(str(constraint))
         for i, sent_idx in enumerate(range(len(results['predictions']))):
             out = ' '.join(
                  [token for token in results['predictions'][sent_idx] if token != END_SYMBOL and token != START_SYMBOL and token != '@@PADDING@@'])
             if i < len(results['predictions']):
                 out = f'{out}\n'
             output_file.open('a').write(out)
-    (output_file.parent / 'constraint.txt').open('w').write('\n'.join(write_constraints))
+    if len(write_constraints) > 0:
+        (output_file.parent / 'constraint.txt').open('w').write('\n'.join(write_constraints))
